refactor: Move IngesterQueryRequest to generated_types
