+++ ingester/src/handler.rs
@@ -13,13 +13,13 @@ use crate::{
 };
 use async_trait::async_trait;
 use backoff::BackoffConfig;
-use data_types2::{IngesterQueryRequest, KafkaPartition, KafkaTopic, Sequencer};
-
+use data_types2::{KafkaPartition, KafkaTopic, Sequencer};
 use futures::{
     future::{BoxFuture, Shared},
     stream::FuturesUnordered,
     FutureExt, StreamExt, TryFutureExt,
 };
+use generated_types::ingester::IngesterQueryRequest;
 use iox_catalog::interface::Catalog;
 use iox_time::SystemProvider;
 use object_store::DynObjectStore;
